move func to netlink utils
